move chainbase's scope_exit and scope_fail to separate header
